+++ src/lists.js
@@ -1,3 +1,13 @@
+/*
+ * This file contains the code for working with scrollable lists.
+ *
+ * Copyright (c) 2010 - 2020 jsPlumb (hello@jsplumbtoolkit.com)
+ *
+ * https://jsplumbtoolkit.com
+ * https://github.com/jsplumb/jsplumb
+ *
+ * Dual licensed under the MIT and GPL2 licenses.
+ */
 ;(function() {
 
     var DEFAULT_OPTIONS = {
@@ -11,11 +21,12 @@
 
     var root = this;
 
-    var ListManager = function(jsPlumbInstance) {
+    var ListManager = function(jsPlumbInstance, params) {
 
         this.count = 0;
         this.instance = jsPlumbInstance;
         this.lists = {};
+        this.options = params || {};
 
         this.instance.addList = function(el, options) {
             return this.listManager.addList(el, options);
@@ -56,6 +67,7 @@
 
         addList : function(el, options) {
             var dp = this.instance.extend({}, DEFAULT_OPTIONS);
+            this.instance.extend(dp, this.options);
             options = this.instance.extend(dp,  options || {});
             var id = [this.instance.getInstanceIndex(), this.count++].join("_");
             this.lists[id] = new List(this.instance, el, options, id);
@@ -161,7 +173,7 @@
                     }
                 }
                 //
-                else if (children[i].offsetTop > el.scrollTop + el.offsetHeight) {
+                else if (children[i].offsetTop + children[i].offsetHeight > el.scrollTop + el.offsetHeight) {
                     if (!children[i]._jsPlumbProxies) {
                         children[i]._jsPlumbProxies = children[i]._jsPlumbProxies || [];